Improve goActions test
This merge request fixes up a little oversight by me. I will merge this as soon as the pipeline passes.

------------------

Some lines verified that nothing is selected when navigating to a
folder that was not acted on yet. These verifications didn't test
anything meaningful because the folder in question was empty.

This commit adds a file and a folder to the test folder so that
testing if nothing is selected means something.