Make details column text readable
When not using the full row highlight, the text of non-name columns
in details view mode was wrongly colored in a way that pretended
that the full row selection highlight was active.

When it isn't active we use the normal color that we generally use
when the additional information is not within the selection
highlight.