Fixed build-time deprecation warning about KStandardGuiItem::yes()
While building dolphin, a warning comes up about KStandardGuiItem::yes()
being deprecated, and that another action verb should be using instead.
This commit fixes the warning by replacing KStandardGuiItem::yes() with
a KGuiItem constructor in the file src/views/dolphinview.cpp. 

The icon for the "Rename and Hide" button has also been changed from a 
check mark (dialog-ok) to an eye with a cross (view-hidden) to make the 
button look more unique and grab the attention of the user.