Reveal the Backspace shortcut in the UI
Since the Back action already has two standard shortcuts, a third one appended
to the list will be invisible in the UI, which shows only the first two.  This
leaves users unable to reassign it.

We therefore make it the first shortcut in the list, so that the hidden third
entry will be the dedicated Back button, which users are far less likely to
want reassigned.

BUG: 371130