Revert "Adapt autotest to new expected "Space" key behaviour"
This reverts commit df5cf6fe851021a73159fdd3487fe629c1ea8376.

The autotest was orginally changed to adapt to new behaviour when
the Space key is pressed. This change used to be implemented in
KItemListController which meant that the KItemListControllerTest
needed changing, but this is no longer the case.
Instead the new behaviour when the Space key is pressed is now
implemented further up the hierarchy, so the KItemListController
change could be reverted in
7b6a67e520c04f56b4b05fa26b252177398df6df and as such we can also
revert this autotest change.