KF6 Dolphin depends on kio-extras, for kcm_trash embedded in settings
When kio-extras is not installed (or kcm_trash.so exists, but
is outdated) then opening the "Configure Dolphin..." settings dialog
will crash. It tries to load kcm_trash with its old KCModule plugin
construction signature and gets a fatal symbol mismatch.

kcm_trash was recently moved from kio to kio-extras and is
unconditionally loaded in non-Windows environments, so that
dependency should be required in the metadata for kdesrc-build.