Never show status bar context menu away from status bar
The status bar context menu was hard-coded to always appear at the
cursor position. However, context menus can also be triggered by
keyboard, for example with the Menu key, in which case it makes no
sense to show the context menu at whatever random position the
mouse cursor currently is.

Instead invoke the context menu in the middle of the status bar
when it is opened by keyboard.