Fix accessibility ancestor tree for folders panel
This is a followup to 549fad2daeeccac53b88b4777dcc9effbc2110e5.
That previous commit made sure that the accessibility tree for
Dolphin's main view was complete even though the view had no
parent by explicitly setting an accessible parent.

The folders panel also has a view though and that previous commit
did not contain an explicit call to also set an accessible parent
for it. This commit rectifies that.

Fixes a sanity check & crash in the debug build.
Belongs to Dolphin issue #47.