Resort directory size count after refreshing
After refreshing the view when
size count is selected as the sortRole,
count is 0 at first. When the actual count is loaded,
the sorting is done according to the count being 0. 
This can break the sort order and cause view and model
to be out of sync. 

Making sure we always resort all items
when the directory size/item count is changed fixes this

BUG:473999