Avoid sorting too frequently
d98037745fe6b5efbe9b145da7d20fa2f731b6a6 changed the time from 500 ms to
50 ms. This commit changes it to 100 ms.

Information relevant for sorting might change repeatedly. Prior to this
commit here we would resort within 50 ms of sorting being requested. If a
lot of resorts would be requested in a short time frame, this could lead
to the item order changing within the view up to 20 times a second which
would lead to a lot of unnecessary movement and make it impossible to
read even file names during the repeated sorting.

100 ms is half as bad in that regard. Bigger values might be even better
    but it is a trade-off.