feat: add option to disable indexer usage for search
Adds an option to allow the user to disable using the indexer for search
inside Dolphin, independent of the global setting in the system settings.

This is useful for users who want to use the indexer for other
applications and situations, but not for Dolphin.

An example of this is when the user wants to use the indexer for
searching in KRunner, but not in Dolphin.