Start autoActivationTimer only if hovering over a directory
Before starting autoActivationTimer, check that we're hovering the item on top of a directory.
If we don't check for it, the the autoActivationTimer will try to open the hovered item 
in it's default application, which can be distracting and break the actual action
the user was trying to do, like moving the file to a directory. 


BUG:479960