Add branding colors for Flathub
This is needed or they definitely won't feature us.

The light blue color is chosen because it is vibrant and fits the
association with "Dolphin". The very dark color is chosen to still have
a good contrast with the Dolphin icon even though it is already quite
dark at the top left corner.

For more information see:
https://docs.flathub.org/blog/introducing-app-brand-colors/