Revert "DragAndDropHelper::updateDropAction: use StatJob for remote URLs"
This reverts commit dc149ec5e52f52c514cf362603d05ba8eea506b8.

This prevents a crash. One issue identified is that the commit that
I am reverting here accesses a QDropEvent at a moment in time in
which it might have already been deleted. We cannot check if it
exists by that time because we do not control its lifetime and it
is not a QObject.