Add label to button for closing the admin bar
The exact effect of the button is not perfectly clear without a
label because it could theoretically also have no effect other than
dismissing the bar (which is not the case). This situation is
clarified through this added label.

The tooltip of the button is adapted to complement the new button
label.