ContextMenuSettingsPage: Stop forcing focus onto search field
Prior to this commit when switching between categories in Dolphin's
settings dialog by keyboard, the focus would suddenly switch away
from the category chooser when going over the "Context Menu"
category. This meant that one couldn't move to the Trash category
in a straightforward way but had to fight the focus change forced
by Dolphin.

This commit removes that forced focus change so the category can
now be changed by keyboard as expected.